メンタルヘルスチェック結果をコンポーネント化対応
